+++ src/lib/components/GlobalFooter.svelte
@@ -18,11 +18,15 @@
 	class="page-x-padding pt-4 pb-4 border-t border-white/10 text-11px flex flex-wrap gap-x-6 gap-y-2 justify-between"
 >
 	<!-- <span>Stem Shovel</span> -->
-	<div class="opacity-70">
-		Built by <a
-			class="inline hover-underline underline-offset-4 hover-text-maximumYellow"
-			href="https://www.lightningjar.com">⚡️ Lightning Jar</a
-		>
+	<div class="flex flex-wrap gap-2">
+		<div>Copyright 2026 Lightning Jar. All rights reserved.</div>
+		<div class="opacity-70">|</div>
+		<div class="opacity-70">
+			Built by <a
+				class="inline hover-underline underline-offset-4 hover-text-maximumYellow"
+				href="https://www.lightningjar.com">⚡️ Lightning Jar</a
+			>
+		</div>
 	</div>
 
 	<div class="flex flex-wrap items-center gap-x-6 gap-y-2">
@@ -58,7 +62,7 @@
 		<a
 			class="underline underline-offset-4 hover:text-maximumYellow opacity-70 underline underline-offset-4 hover-text-maximumYellow"
 			href="https://www.youtube.com/watch?v=dQw4w9WgXcQ"
-			title="Click in case of emergency">For Your Eyes Only</a
+			title="Click in case of emergency">RR</a
 		>
 	</div>
 </footer>